Task 1 implemented.
All tests are being passed. Some accessor methods were implemented.
getSystemTime() which return current sim ssytem time.
getNumProcesses() which returns the total number of processes that will be in the sim.
isCpuIdle() returns true if cpu is IDLE and false if it isn't.
Lastly getRunningProcessName() which returns "IDLE" if cpu is idle and the process name of the current cpu process id if it isn't.