Task 4 implemented.
All tests are being passed.
Method checkProcessFinished() implemented.
This method does nothing if cpu is idle or if the process used time is less
than its service time.
Otherwise it will record its endtime as current systemTIme.
Mark the process as done.
And set the cpu to be idle again.